requirements.txt
Flask
PyMuPDF
qrcode
Pillow
num2words
gunicorn
